+++ CHANGES.md
@@ -1,6 +1,7 @@
 v4.1.0
 ------
 
+- Install the Bubblewrap sandboxing tool in all distributions.
 - Ensure that X11 is available in the containers so that the
   OCaml Graphics module is available (#8 via @kit-ty-kate)
 - Add concept of a "Tier 1" and "Tier 2" distro so that we can